[#910] fix conversion and sign-conversion errors
Except for tests:
- remove -Wno-conversion and -Wno-sign-conversion flags
- when error is caused by external header include, use GCC diagnostic
pragma to narrow down warning ignore to the offending header only
- use explicit static_cast<X> when necessary
- if conversion "error" (eg: sign change) would lead into corrupted
state (eg. index arithmetic) an exception is thrown. Otherwise casting
is done without check (eg. rviz marker.id)

+++ src/tools/visualization/autoware_rviz_plugins/src/object_detection/bounding_box_array_display.cpp
@@ -87,7 +87,7 @@ void BoundingBoxArrayDisplay::processMessage(
     const auto marker_ptr = get_marker(msg->boxes[idx]);
     marker_ptr->ns = "bounding_box";
     marker_ptr->header = msg->header;
-    marker_ptr->id = idx;
+    marker_ptr->id = static_cast<int>(idx);
     m_marker_common->addMessage(marker_ptr);
   }
 }
